move Contacts fetch to App.js
