+++ client_side/src/pages/Contact.js
@@ -7,35 +7,14 @@ Type: Adapted from
 Source: https://github.com/iamshaunjp/MERN-Auth-Tutorial
 */
 
-import { useEffect } from "react"
 import { useContactsContext } from "../hooks/useContactsContext";
-import { useAuthContext } from "../hooks/useAuthContext"
 
 // components
 import ContactDetails from "../components/ContactDetails"
 import ContactForm from "../components/ContactForm"
 
 const Contact = () => {
-    const { contacts, dispatch } = useContactsContext()
-    const { user } = useAuthContext()
-
-    useEffect(() => {
-        const fetchContacts = async () => {
-            const response = await fetch('/contacts', {
-                headers: { 'Authorization': `Bearer ${user.token}` },
-            })
-            const json = await response.json()
-
-            if (response.ok) {
-                dispatch({ type: 'SET_CONTACTS', payload: json })
-            }
-        }
-
-        if (user) {
-            fetchContacts()
-        }
-        fetchContacts()
-    }, [dispatch, user])
+    const { contacts } = useContactsContext()
 
     return (
         <div className="home">
